+++ whatsnew.md
@@ -15,8 +15,10 @@
 - STARS
   - Added support for "Create interfacility VFR flight plan from active local track"
   - Fixed bug where WX3 would never have any contents
+  - GI and ATIS text in the SSA list is managed by the server (so updates are shared among controllers)
 - Facility engineering:
   - Support "at or above/at or below" speed restrictions
   - Removed "speed_restriction" from departure routes; this was unused/redundant
   - Added "thick_20nm_range_ring" STARS adaptation
   - Added "flight_type" for overflights so aircraft flight types can be specified as "departure" or "arrival" (rather than always being "overflight")
+  - Added "flash_on_atis_update" adaptation for STARS SSA lists